Users endpoint. Added two proxy objects to control which userResource/userCollectionResource manage the request depending on version number
